Added
I have added 2 excels with the results. In the excel file, you will find different combinations of δ and ψ. The name of each excel file corresponds to a different approach. For example, ...O3T150.csv means that we run the code with O [1:3] and T [1:150]. In addition, a script was added to plot the results based on the csvs.